chore: remove hosted deployment type
