+++ app/apphandlers/processorAppHandler.go
@@ -201,7 +201,7 @@ func (processor *ProcessorApp) StartRudderCore(ctx context.Context, options *app
 	case deployment.MultiTenantType:
 		pkgLogger.Info("using ETCD Based Dynamic Cluster Manager")
 		modeProvider = state.NewETCDDynamicProvider()
-	case deployment.HostedType, deployment.DedicatedType:
+	case deployment.DedicatedType:
 		// FIXME: hacky way to determine servermode
 		pkgLogger.Info("using Static Cluster Manager")
 		if enableProcessor && enableRouter {
